refactor(sync): share visibility predicates across core and server
